requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
behave==1.2.6
apispec-webframeworks==0.5.2